+++ libsolidity/analysis/ReferencesResolver.h
@@ -91,6 +91,9 @@ private:
 	/// Adds a new error to the list of errors.
 	void declarationError(langutil::SourceLocation const& _location, std::string const& _description);
 
+	/// Adds a new error to the list of errors.
+	void declarationError(langutil::SourceLocation const& _location, langutil::SecondarySourceLocation const& _ssl, std::string const& _description);
+
 	/// Adds a new error to the list of errors and throws to abort reference resolving.
 	void fatalDeclarationError(langutil::SourceLocation const& _location, std::string const& _description);